# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Consumer Product Recalls API",
    "description": "Open, read-only API for US consumer product recalls from five agencies (CPSC, FDA, USDA, NHTSA, USCG). No key, GET-only, cursor-paginated. A few things worth knowing up front: a recall's `is_active` flag is `null` (not false) for CPSC and NHTSA, which don't track an open/closed status; `classification` uses each agency's own severity scale and can't be compared across agencies; UPC search matches a whole recall, not an individual product; and search is exact — a typo finds nothing.\n\n**Common lookups:** by product name → `GET /products/search?q=`; by **UPC barcode** → `GET /products/search?upc=`; vehicle or boat by identifier → `GET /products/search?model=` / `?hin=`; a single recall → `GET /recalls/{source}/{recall_id}`.",
    "version": "0.1.1"
  },
  "paths": {
...
    "/products/search": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Search recalled products by keyword (full-text) or exact identifier.",
        "description": "Search recalled products — \"is my product recalled?\". Keyword (`q`) is a full-text search over product name, description, recall title, and firm, ranked by relevance (`rank`); whole words and prefixes only, no fuzzy or typo search. `hin` and `model` are exact-match lookups. `upc` matches at the recall level, so each hit carries `upc_is_recall_level: true` — a miss means no recall lists that UPC, not that the product is safe. Supply at least one of `q`, `hin`, `model`, or `upc` (otherwise 422); if several are given, `q` wins, then `hin`/`model`, then `upc`. `source` narrows any of these and accepts multiple values (repeat or comma-separate).",
        "operationId": "search_products_products_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Keywords to search for.",
              "title": "Q"
            },
            "description": "Keywords to search for."
          },
          {
            "name": "hin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact Hull ID (USCG boats).",
              "title": "Hin"
            },
            "description": "Exact Hull ID (USCG boats)."
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact product model.",
              "title": "Model"
            },
            "description": "Exact product model."
          },
          {
            "name": "upc",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "description": "Look up recalled products by 12-digit UPC barcode. Matches at the recall level — UPC data is sparse and comes only from CPSC (about 5% of CPSC recalls; absent for FDA, USDA, NHTSA, and USCG). A miss means no recall lists that UPC, **not** that the product is safe.",
              "examples": [
                "012345678905"
              ],
              "title": "Upc"
            },
            "description": "Look up recalled products by 12-digit UPC barcode. Matches at the recall level — UPC data is sparse and comes only from CPSC (about 5% of CPSC recalls; absent for FDA, USDA, NHTSA, and USCG). A miss means no recall lists that UPC, **not** that the product is safe."
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Source"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional agency filter; repeat/comma-separate to match any of several.",
              "title": "Source"
            },
            "description": "Optional agency filter; repeat/comma-separate to match any of several."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size (max 100).",
              "default": 25,
              "title": "Limit"
            },
            "description": "Page size (max 100)."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a prior next_cursor.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a prior next_cursor."
          },
          {
            "name": "with_total",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also return the total count (an extra query).",
              "default": false,
              "title": "With Total"
            },
            "description": "Also return the total count (an extra query)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProductSearchHit_"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/firms/{firm_id}": {
      "get": {
        "tags": [
          "firms"
        ],
        "summary": "Fetch a firm profile with its agency registration records.",
        "description": "A single firm, merged across agencies — so a maker that appears under several (e.g. Honda under NHTSA and USCG) shows up as one profile. Includes recall counts, the per-agency breakdown (`recalls_by_source`), name variants, and agency registration records: `firm_usda_attributes` (USDA establishments), `firm_uscg_attributes` (USCG boat builders), and `firm_fda_attributes` (FDA-registered firms). Each set has a different shape and any may be empty; CPSC and NHTSA contribute none. `first_recall_at` and `last_recall_at` are null for a firm with no matched recalls. The id is opaque — get it from a recall's `firms[].firm_id`.",
        "operationId": "get_firm_firms__firm_id__get",
        "parameters": [
          {
            "name": "firm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 32,
              "maxLength": 32,
              "pattern": "^[0-9a-f]{32}$",
              "description": "Opaque canonical firm id (md5 cluster id).",
              "title": "Firm Id"
            },
            "description": "Opaque canonical firm id (md5 cluster id)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirmProfile"
                }
              }
            }
          },
          "404": {
            "description": "ResourceNotFound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "examples": [
              "not_found"
            ]
          },
          "detail": {
            "title": "Detail",
            "examples": [
              "No recall found for CPSC/24-001."
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "examples": [
              "b1d9c6f2-3a1e-4c7e-9f0a-7d2c1e5b8a40"
            ]
          }
        },
        "type": "object",
        "required": [
          "type",
          "detail",
          "request_id"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
      "FdaAttributes": {
        "properties": {
          "firm_fei_num": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Firm Fei Num"
          },
          "firm_legal_nam": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firm Legal Nam"
          },
          "firm_city_nam": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firm City Nam"
          },
          "firm_state_cd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firm State Cd"
          },
          "firm_state_prvnc_nam": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firm State Prvnc Nam"
          },
          "firm_country_nam": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firm Country Nam"
          },
          "firm_postal_cd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firm Postal Cd"
          },
          "firm_line1_adr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firm Line1 Adr"
          },
          "firm_line2_adr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firm Line2 Adr"
          },
          "firm_surviving_nam": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firm Surviving Nam"
          },
          "firm_surviving_fei": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firm Surviving Fei"
          }
        },
        "type": "object",
        "required": [
          "firm_fei_num"
        ],
        "title": "FdaAttributes",
        "description": "An FDA FEI firm row (firm_fda_attributes, join key firm_fei_num cast to text)."
      },
...
      "FirmProfile": {
        "properties": {
          "firm_id": {
            "type": "string",
            "title": "Firm Id",
            "description": "Synthetic primary key of the canonical firm (a cross-source cluster). Derived from the firm-resolution crosswalk, falling back to md5(normalized name). One row per canonical firm across all agencies. Sources: derived (all five contribute names)."
          },
          "canonical_name": {
            "type": "string",
            "title": "Canonical Name",
            "description": "Human-readable display name of the canonical firm — the cluster's resolved name, or the representative raw firm name when unclustered. Not an authoritative legal name. Sources: all five.",
            "examples": [
              "Acme Foods Inc"
            ]
          },
          "normalized_name": {
            "type": "string",
            "title": "Normalized Name",
            "description": "Upper-cased, whitespace-trimmed form of the firm's representative name for case-insensitive lookup. NOT a unique key (firm_id is the key). Sources: all five."
          },
          "observed_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Observed Names",
            "description": "JSONB array of all distinct raw firm-name surface forms (across sources and spellings) that map to this canonical firm — the provenance/audit trail of names collapsed together. Always >=1 element. Sources: all five."
          },
          "observed_company_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Observed Company Ids",
            "description": "JSONB array of distinct structured firm identifiers observed for this firm: FDA FEI numbers, USDA FSIS establishment numbers, and USCG MICs. Also the join key to the three sidecars. Sources: FDA, USDA, USCG (empty for firms seen only via CPSC/NHTSA, which carry no usable firm id)."
          },
          "alternate_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alternate Names",
            "description": "JSONB array of brand/DBA surface-form aliases (e.g. 'John Deere' for 'Deere & Company (John Deere)'), derived by the firm-resolution step for search and fuzzy matching. Empty when the firm has no aliases. Sources: derived (not a per-agency field)."
          },
          "total_recalls": {
            "type": "integer",
            "title": "Total Recalls",
            "description": "Total distinct recalls this firm is linked to, across all sources it appears in (a firm in multiple roles on one recall counts once). Always present. Sources: all five.",
            "default": 0
          },
          "active_recalls": {
            "type": "integer",
            "title": "Active Recalls",
            "description": "Count of this firm's distinct currently-active recalls. Only FDA, USDA, and USCG recalls can be active; CPSC and NHTSA have no lifecycle (is_active null) and never count. Always present (0 if none). Sources counted: FDA, USDA, USCG.",
            "default": 0
          },
          "first_recall_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Recall At",
            "description": "Earliest recall date for this firm — min of coalesce(announced_at, published_at): the announce/initiation date, falling back to published_at where announced_at is null. Null only for a firm with no linked recall. Sources: all five."
          },
          "last_recall_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Recall At",
            "description": "Most recent recall date for this firm — max of coalesce(announced_at, published_at) (same announce-date basis as first_recall_at). Null only for a firm with no linked recall. Sources: all five."
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "description": "Distinct roles this firm has played across its recalls: manufacturer, importer, distributor (CPSC), establishment (FDA/USDA), filer/manufacturer (NHTSA), manufacturer (USCG). Sources: all five."
          },
          "recalls_by_source": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Recalls By Source",
            "description": "JSONB object mapping source -> distinct recall count for this firm (e.g. {'NHTSA': 12, 'CPSC': 3}). Only sources where the firm has >=1 recall appear as keys, and the values sum to total_recalls. Sources: all five.",
            "examples": [
              {
                "FDA": 2,
                "USDA": 1
              }
            ]
          },
          "distinct_products": {
            "type": "integer",
            "title": "Distinct Products",
            "description": "Total distinct recalled-product rows across all recalls this firm is associated with, in any role (a per-firm footprint, NOT a global distinct — a product on a multi-firm recall is counted under each firm). Never null. Sources: all five.",
            "default": 0
          },
          "firm_usda_attributes": {
            "items": {
              "$ref": "#/components/schemas/UsdaEstablishment"
            },
            "type": "array",
            "title": "Firm Usda Attributes",
            "description": "USDA FSIS establishment attributes — a JSON array of one block per matched FSIS establishment number (name, address, regulatory metadata, grant dates). Empty for non-USDA firms. Sources: USDA only."
          },
          "firm_uscg_attributes": {
            "items": {
              "$ref": "#/components/schemas/UscgManufacturer"
            },
            "type": "array",
            "title": "Firm Uscg Attributes",
            "description": "USCG boat-manufacturer directory attributes — a JSON array of one block per USCG Manufacturer Identification Code (MIC) the firm is registered under (company name, address, status, succession lineage, MIC-recycle flags). Empty for non-USCG firms. Sources: USCG only."
          },
          "firm_fda_attributes": {
            "items": {
              "$ref": "#/components/schemas/FdaAttributes"
            },
            "type": "array",
            "title": "Firm Fda Attributes",
            "description": "FDA establishment attributes — a JSON array of one block per FDA FEI the firm clusters under (legal name, address at time of recall, firm-succession signal). Empty for non-FDA firms. Sources: FDA only."
          }
        },
        "type": "object",
        "required": [
          "firm_id",
          "canonical_name",
          "normalized_name"
        ],
        "title": "FirmProfile",
        "description": "One canonical (cross-source) firm: identity, aliases, recall stats, and 3 sidecars."
      },
...
      "Page_ProductSearchHit_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductSearchHit"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "examples": [
              "eyJ2IjpbIjIwMjYtMDEtMDEiLCJhYmMiXX0"
            ]
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "examples": [
              25
            ]
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total",
            "description": "Only when with_total=true.",
            "examples": [
              null
            ]
          }
        },
        "type": "object",
        "required": [
          "items",
          "limit"
        ],
        "title": "Page[ProductSearchHit]"
      },
...
      "Source": {
        "type": "string",
        "enum": [
          "CPSC",
          "FDA",
          "USDA",
          "NHTSA",
          "USCG"
        ],
        "title": "Source",
        "description": "The one closed cross-source domain (uppercase). Free 422 on a bad path/query value."
      },
...
      "UsdaEstablishment": {
        "properties": {
          "establishment_id": {
            "type": "string",
            "title": "Establishment Id"
          },
          "establishment_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Establishment Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "zip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip"
          },
          "county": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "County"
          },
          "fips_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fips Code"
          },
          "geolocation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geolocation"
          },
          "latest_mpi_active_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Mpi Active Date"
          },
          "grant_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Date"
          },
          "status_regulated_est": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Regulated Est"
          },
          "size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "circuit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Circuit"
          },
          "activities": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activities"
          },
          "dbas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dbas"
          }
        },
        "type": "object",
        "required": [
          "establishment_id"
        ],
        "title": "UsdaEstablishment",
        "description": "A USDA/FSIS establishment row (firm_usda_attributes, join key establishment_id)."
      }
    }
  }
}